Remove validate_funding_method_eligibility
